js/xpconnect/tests/mochitest/test_crosscompartment_weakmap.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Cross-Compartment DOM WeakMaps</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<script type="application/javascript">

var my_map = new WeakMap();

function setup() {
  var item = window.frames[0].document.querySelector("body");

  my_map.set(item, "success_string");
}

function runTest() {
  setup();
  SpecialPowers.forceGC();
  SpecialPowers.forceCC();
  SpecialPowers.forceGC();
  SpecialPowers.forceCC();
  var item = window.frames[0].document.querySelector("body");
  is(my_map.get(item), "success_string", "Preserve reflectors used cross-compartment as weak map keys.");
}

</script>
<iframe src="file_crosscompartment_weakmap.html" onload="runTest()"></iframe>


</pre>
</body>
</html>
